# blog-gin/docs/swagger.yaml
definitions:
...
  dto.UserApi:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.UserApi'
        type: array
      created_at:
        description: 创建时间
        type: integer
      id:
        description: 主键id
        type: integer
      method:
        description: api请求方法
        type: string
      name:
        description: api名称
        type: string
      parent_id:
        description: 父id
        type: integer
      path:
        description: api路径
        type: string
      updated_at:
        description: 更新时间
        type: integer
    type: object
  dto.UserApisResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.UserApi'
        type: array
    type: object
...
  dto.UserInfoReq:
    properties:
      avatar:
        description: 头像
        type: string
      intro:
        description: 简介
        type: string
      nickname:
        description: 昵称
        type: string
      website:
        description: 网站
        type: string
    type: object
...
  dto.UserLoginHistoryQuery:
    properties:
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      sorts:
        description: 排序
        items:
          type: string
        type: array
    type: object
  dto.UserMenu:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.UserMenu'
        type: array
      component:
        description: Layout组件
        type: string
      created_at:
        description: 创建时间
        type: integer
      id:
        description: 主键
        type: integer
      meta:
        allOf:
          - $ref: '#/definitions/dto.UserMenuMeta'
        description: meta配置
      name:
        description: 路由名字
        type: string
      parent_id:
        description: 父id
        type: integer
      path:
        description: 路由地址
        type: string
      redirect:
        description: 路由重定向
        type: string
      updated_at:
        description: 更新时间
        type: integer
    type: object
  dto.UserMenuMeta:
    properties:
      affix:
        type: boolean
      alwaysShow:
        type: boolean
      breadcrumb:
        type: boolean
      hidden:
        type: boolean
      icon:
        type: string
      keepAlive:
        type: boolean
      title:
        type: string
    type: object
  dto.UserMenusResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.UserMenu'
        type: array
    type: object
  dto.UserRole:
    properties:
      id:
        description: 主键id
        type: integer
      parent_id:
        description: 父id
        type: integer
      role_comment:
        description: 角色备注
        type: string
      role_label:
        description: 角色标签
        type: string
      role_name:
        description: 角色名
        type: string
    type: object
  dto.UserRoleLabel:
    properties:
      role_comment:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
    type: object
  dto.UserRolesResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.UserRole'
        type: array
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.EmptyReq:
    type: object
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.EmptyResp:
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.PageResp:
    properties:
      list: { }
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.UserInfoResp:
    properties:
      avatar:
        description: 用户头像
        type: string
      created_at:
        type: integer
      email:
        description: 用户邮箱
        type: string
      intro:
        description: 简介
        type: string
      ip_address:
        description: ip host
        type: string
      ip_source:
        description: ip 源
        type: string
      login_type:
        description: 登录方式
        type: string
      nickname:
        description: 用户昵称
        type: string
      perms:
        items:
          $ref: '#/definitions/dto.UserApi'
        type: array
      phone:
        description: 用户手机号
        type: string
      roles:
        items:
          $ref: '#/definitions/dto.UserRoleLabel'
        type: array
      status:
        description: 状态
        type: integer
      updated_at:
        type: integer
      user_id:
        description: 用户id
        type: string
      username:
        description: 用户名
        type: string
      website:
        description: 网站
        type: string
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: { }
      message:
        type: string
      trace_id:
        type: string
    type: object
host: localhost:9090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: https://www.swagger.io/support/
  description: APP接口鉴权方式：token采用自加密token的方式，自加密token规则：sha256(${x-timestamp}${base_secret})
    生成的加密token，x-timestamp 时间戳需要保持10分钟内
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Swagger Example API
  version: 1.0.0
paths:
...
  /admin_api/v1/user/get_user_apis:
    get:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.EmptyReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/dto.UserApisResp'
                type: object
      summary: '"获取用户接口权限"'
      tags:
        - User
  /admin_api/v1/user/get_user_info:
    get:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.EmptyReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.UserInfoResp'
                type: object
      summary: '"获取用户信息"'
      tags:
        - User
  /admin_api/v1/user/get_user_login_history_list:
    post:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.UserLoginHistoryQuery'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.PageResp'
                type: object
      summary: '"查询用户登录历史"'
      tags:
        - User
  /admin_api/v1/user/get_user_menus:
    get:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.EmptyReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/dto.UserMenusResp'
                type: object
      summary: '"获取用户菜单权限"'
      tags:
        - User
  /admin_api/v1/user/get_user_roles:
    get:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.EmptyReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/dto.UserRolesResp'
                type: object
      summary: '"获取用户角色"'
      tags:
        - User
  /admin_api/v1/user/update_user_info:
    post:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.UserInfoReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.EmptyResp'
                type: object
      summary: '"修改用户信息"'
      tags:
        - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"